rearrange some types
