+++ app/context/CurrentLayoutContext/actions.ts
@@ -14,14 +14,8 @@ import { MosaicNode, MosaicPath } from "react-mosaic-component";
 
 import { GlobalVariables } from "@foxglove/studio-base/hooks/useGlobalVariables";
 import { LinkedGlobalVariables } from "@foxglove/studio-base/panels/ThreeDimensionalViz/Interactions/useLinkedGlobalVariables";
-import { Dispatcher } from "@foxglove/studio-base/reducers";
 import { TabLocation } from "@foxglove/studio-base/types/layouts";
 import {
-  CreateTabPanelPayload,
-  LoadLayoutPayload,
-  ChangePanelLayoutPayload,
-  SaveConfigsPayload,
-  SaveFullConfigPayload,
   UserNodes,
   PlaybackConfig,
   SavedProps,
@@ -43,6 +37,41 @@ export type PanelsState = {
   playbackConfig: PlaybackConfig;
   version?: number;
 };
+
+export type ConfigsPayload = {
+  id: string;
+  // if you set override to true, existing config will be completely overriden by new passed in config
+  override?: boolean;
+  config: PanelConfig;
+  defaultConfig?: PanelConfig;
+};
+export type ChangePanelLayoutPayload = {
+  layout?: MosaicNode<string>;
+  trimConfigById?: boolean;
+};
+export type SaveConfigsPayload = {
+  configs: ConfigsPayload[];
+};
+
+type PerPanelFunc<Config> = (arg0: Config) => Config;
+export type SaveFullConfigPayload = {
+  panelType: string;
+  perPanelFunc: PerPanelFunc<PanelConfig>;
+};
+
+export type CreateTabPanelPayload = {
+  idToReplace?: string;
+  layout: MosaicNode<string>;
+  idsToRemove: string[];
+  singleTab: boolean;
+};
+
+export type LoadLayoutPayload = Partial<Omit<PanelsState, "id" | "name">>;
+
+export type UpdatePanelConfig<Config> = (
+  panelType: string,
+  perPanelFunc: PerPanelFunc<Config>,
+) => void;
 
 export enum PANELS_ACTION_TYPES {
   CHANGE_PANEL_LAYOUT = "CHANGE_PANEL_LAYOUT",
@@ -78,47 +107,12 @@ export type CHANGE_PANEL_LAYOUT = {
 };
 export type LOAD_LAYOUT = { type: "LOAD_LAYOUT"; payload: LoadLayoutPayload };
 
-export const savePanelConfigs =
-  (payload: SaveConfigsPayload): Dispatcher<SAVE_PANEL_CONFIGS> =>
-  (dispatch) => {
-    return dispatch({ type: PANELS_ACTION_TYPES.SAVE_PANEL_CONFIGS, payload });
-  };
-
-export const saveFullPanelConfig =
-  (payload: SaveFullConfigPayload): Dispatcher<SAVE_FULL_PANEL_CONFIG> =>
-  (dispatch) => {
-    return dispatch({ type: PANELS_ACTION_TYPES.SAVE_FULL_PANEL_CONFIG, payload });
-  };
-
-export const createTabPanel = (payload: CreateTabPanelPayload): CREATE_TAB_PANEL => ({
-  type: PANELS_ACTION_TYPES.CREATE_TAB_PANEL,
-  payload,
-});
-
-export const changePanelLayout =
-  (payload: ChangePanelLayoutPayload): Dispatcher<CHANGE_PANEL_LAYOUT> =>
-  (dispatch) => {
-    return dispatch({ type: PANELS_ACTION_TYPES.CHANGE_PANEL_LAYOUT, payload });
-  };
-
-export const loadLayout =
-  (payload: LoadLayoutPayload): Dispatcher<LOAD_LAYOUT> =>
-  (dispatch) => {
-    return dispatch({ type: PANELS_ACTION_TYPES.LOAD_LAYOUT, payload });
-  };
-
 export type OVERWRITE_GLOBAL_DATA = {
   type: "OVERWRITE_GLOBAL_DATA";
   payload: {
     [key: string]: unknown;
   };
 };
-export const overwriteGlobalVariables = (payload: {
-  [key: string]: unknown;
-}): OVERWRITE_GLOBAL_DATA => ({
-  type: PANELS_ACTION_TYPES.OVERWRITE_GLOBAL_DATA,
-  payload,
-});
 
 export type SET_GLOBAL_DATA = {
   type: "SET_GLOBAL_DATA";
@@ -126,33 +120,15 @@ export type SET_GLOBAL_DATA = {
     [key: string]: unknown;
   };
 };
-export const setGlobalVariables = (payload: { [key: string]: unknown }): SET_GLOBAL_DATA => ({
-  type: PANELS_ACTION_TYPES.SET_GLOBAL_DATA,
-  payload,
-});
 
 export type SET_STUDIO_NODES = { type: "SET_USER_NODES"; payload: UserNodes };
-export const setUserNodes = (payload: UserNodes): SET_STUDIO_NODES => ({
-  type: PANELS_ACTION_TYPES.SET_USER_NODES,
-  payload,
-});
 
 export type SET_LINKED_GLOBAL_VARIABLES = {
   type: "SET_LINKED_GLOBAL_VARIABLES";
   payload: LinkedGlobalVariables;
 };
-export const setLinkedGlobalVariables = (
-  payload: LinkedGlobalVariables,
-): SET_LINKED_GLOBAL_VARIABLES => ({
-  type: PANELS_ACTION_TYPES.SET_LINKED_GLOBAL_VARIABLES,
-  payload,
-});
 
 export type SET_PLAYBACK_CONFIG = { type: "SET_PLAYBACK_CONFIG"; payload: Partial<PlaybackConfig> };
-export const setPlaybackConfig = (payload: Partial<PlaybackConfig>): SET_PLAYBACK_CONFIG => ({
-  type: PANELS_ACTION_TYPES.SET_PLAYBACK_CONFIG,
-  payload,
-});
 
 export type ClosePanelPayload = {
   tabId?: string;
@@ -160,10 +136,6 @@ export type ClosePanelPayload = {
   path: MosaicPath;
 };
 export type CLOSE_PANEL = { type: "CLOSE_PANEL"; payload: ClosePanelPayload };
-export const closePanel = (payload: ClosePanelPayload): CLOSE_PANEL => ({
-  type: PANELS_ACTION_TYPES.CLOSE_PANEL,
-  payload,
-});
 
 export type SplitPanelPayload = {
   tabId?: string;
@@ -174,10 +146,6 @@ export type SplitPanelPayload = {
   config: PanelConfig;
 };
 export type SPLIT_PANEL = { type: "SPLIT_PANEL"; payload: SplitPanelPayload };
-export const splitPanel = (payload: SplitPanelPayload): SPLIT_PANEL => ({
-  type: PANELS_ACTION_TYPES.SPLIT_PANEL,
-  payload,
-});
 
 export type SwapPanelPayload = {
   tabId?: string;
@@ -189,17 +157,9 @@ export type SwapPanelPayload = {
   relatedConfigs?: SavedProps;
 };
 export type SWAP_PANEL = { type: "SWAP_PANEL"; payload: SwapPanelPayload };
-export const swapPanel = (payload: SwapPanelPayload): SWAP_PANEL => ({
-  type: PANELS_ACTION_TYPES.SWAP_PANEL,
-  payload,
-});
 
 export type MoveTabPayload = { source: TabLocation; target: TabLocation };
 export type MOVE_TAB = { type: "MOVE_TAB"; payload: MoveTabPayload };
-export const moveTab = (payload: MoveTabPayload): MOVE_TAB => ({
-  type: PANELS_ACTION_TYPES.MOVE_TAB,
-  payload,
-});
 
 export type AddPanelPayload = {
   /** id must be formatted as returned by `getPanelIdForType`. This is required as an argument
@@ -212,10 +172,6 @@ export type AddPanelPayload = {
   relatedConfigs?: SavedProps;
 };
 export type ADD_PANEL = { type: "ADD_PANEL"; payload: AddPanelPayload };
-export const addPanel = (payload: AddPanelPayload): ADD_PANEL => ({
-  type: PANELS_ACTION_TYPES.ADD_PANEL,
-  payload,
-});
 
 export type DropPanelPayload = {
   newPanelType: string;
@@ -226,20 +182,12 @@ export type DropPanelPayload = {
   relatedConfigs?: SavedProps;
 };
 export type DROP_PANEL = { type: "DROP_PANEL"; payload: DropPanelPayload };
-export const dropPanel = (payload: DropPanelPayload): DROP_PANEL => ({
-  type: PANELS_ACTION_TYPES.DROP_PANEL,
-  payload,
-});
 
 export type StartDragPayload = {
   path: MosaicPath;
   sourceTabId?: string;
 };
 export type START_DRAG = { type: "START_DRAG"; payload: StartDragPayload };
-export const startDrag = (payload: StartDragPayload): START_DRAG => ({
-  type: PANELS_ACTION_TYPES.START_DRAG,
-  payload,
-});
 
 export type EndDragPayload = {
   originalLayout: MosaicNode<string>;
